(maint) Fix echo task test
Previously the `EchoTask` mock would recieve `:run` and *always* return the expected result. This was masking an issue where the echo task result as defined should return a result that starts with "Hi, ..." but instead always returned "Hello, ...". This commit updates the test to compare the result to be printed to STDOUT.