(maint) Update examples so they are actually functional
Previously the examples did not work. The biggest issue is the relative require issue that is described in MODULES-8605. The updated example will show a way around that by conditionally requiring based on the existence of the task helper source. This commit also fixes the discrepency between parameter name exepected in the example task and it's metadata. Finally it fixes up the test so that it references the right task class name and actually tests that STDOUT has the expected value instead of an incorrect use of Rspec Mock.