Wrap errors in _error key
Previously, the task helper would rescue errors in the `run` method and
simply print their hashified version. This caused Bolt to synthesize its own
`_error` object (simply stating the task failed and its exit code) which
took precedence over the actual underlying cause.

We now properly wrap the result in `_error` so Bolt knows that it should
display that information to the user.