Also serialize Arrays to JSON
The code currently serialize Hash to JSON, but does not do so for
Array, resulting in a serialization using #to_s to be done.  While an
Array of scalars is serialized by #to_s in a String which is a valid
JSON document, Hash are serialized with a Ruby specific-notation, i.e.
`{:foo=>"bar"}` instead of `{"foo": "bar"}`.  As a result, a task which
returns an array of structured results does not emit a processable
output.

Ensure a returned Array is serialized to JSON by the helper so that any
content can be processed as valid JSON data.